--- docs/packages/twoslash/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "abitype/docs-twoslash",
-  "private": true,
-  "devDependencies": {
-    "abitype": "*",
-    "ethers": "*",
-    "@ethersproject/abi": "*"
-  },
-  "dependencies": {}
-}
