--- docs/packages/twoslash/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "jsx": "preserve",
-    "lib": ["DOM", "ESNext"],
-    "module": "esnext",
-    "moduleResolution": "node",
-    "noUnusedLocals": true,
-    "paths": {
-      "~/*": ["src/*"]
-    },
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "strictNullChecks": true,
-    "target": "esnext",
-    "types": ["vite/client", "vitepress"]
-  },
-  "exclude": ["dist", "node_modules"]
-}
